proteccion rutas y cntrller de empleado

+++ app/Http/Controllers/ProductosController.php
@@ -13,8 +13,25 @@ class ProductosController extends Controller
     //
     public function index()
     {
-        $productos = Producto::all();
-        return view('productos.indexProd', compact('productos'));
+        $productos = Producto::simplePaginate(15);
+        $totalUsers = Producto::count();
+        $totalPages = ceil($totalUsers / 15);
+
+        return view('productos.indexProd', compact('productos', 'totalPages'));
+    }
+
+    public function search(Request $request)
+    {
+        $term = $request->input('term');
+
+        $productos = Producto::where('tipoproducto', 'like', "%$term%")
+                              ->orWhere('referencia', 'like', "%$term%")
+                              ->orWhere('descripcion', 'like', "%$term%")
+                              ->orWhere('alto', 'like', "%$term%")
+                              ->orWhere('ancho', 'like', "%$term%")
+                              ->get();
+
+        return response()->json($productos);
     }
 
 
@@ -24,37 +41,6 @@ class ProductosController extends Controller
     }
 
 
-
-
-    // public function store(Request $request)
-    // {
-    //     // Validar los datos del formulario
-    //     $request->validate([
-    //         'tipoproducto' => 'required|string',
-    //         'referencia' => 'required|string',
-    //         'descripcion' => 'required|string',
-    //         'alto' => 'required|string',
-    //         'ancho' => 'required|string',
-    //         'plano' => 'file|mimes:pdf|max:2048',
-    //     ]);
-
-    //     // Procesar la carga del PDF
-    //     $planoPath = $request->file('plano')->store('public/pdfs');
-
-    //     // Crear un nuevo producto en la base de datos
-    //     $producto = new Producto();
-    //     $producto->tipoproducto = $request->tipoproducto;
-    //     $producto->referencia = $request->referencia;
-    //     $producto->descripcion = $request->descripcion;
-    //     $producto->alto = $request->alto;
-    //     $producto->ancho = $request->ancho;
-    //     // Asignar la ruta del archivo al campo 'plano'
-    //     $producto->plano = str_replace('public/', '', $planoPath);
-    //     $producto->save();
-
-    //     // Redireccionar al usuario a la lista de productos con un mensaje de éxito
-    //     return redirect()->route('product.index')->with('success', 'Producto creado correctamente');
-    // }
     public function store(Request $request)
     {
         // Validar los datos del formulario
@@ -110,25 +96,21 @@ class ProductosController extends Controller
     //     return response()->file($pdfUrl);
     // }
     public function showPdf($id)
-{
-    $producto = Producto::find($id);
-    if (!$producto || !$producto->plano) {
-        // Session::flash('error', 'saaaaaaaaa creado exitosamente');
-        return redirect()->back()->with('error', 'Este producto no contiene ningún archivo de soporte pdf por favor edita aquel producto y agregale su respectivo soporte.');
+    {
+        $producto = Producto::find($id);
+        if (!$producto || !$producto->plano) {
+            // Session::flash('error', 'saaaaaaaaa creado exitosamente');
+            return redirect()->back()->with('error', 'Este producto no contiene ningún archivo de soporte pdf por favor edita aquel producto y agregale su respectivo soporte.');
+        }
+
+        $pdfPath = storage_path('app/public/' . $producto->plano);
+
+        if (!file_exists($pdfPath)) {
+            return redirect()->back()->with('error', 'El archivo PDF no existe.');
+        }
+
+        return response()->file($pdfPath);
     }
-
-    $pdfPath = storage_path('app/public/' . $producto->plano);
-
-    if (!file_exists($pdfPath)) {
-        return redirect()->back()->with('error', 'El archivo PDF no existe.');
-    }
-
-    return response()->file($pdfPath);
-}
-
-
-
-
 
 
 
@@ -136,22 +118,6 @@ class ProductosController extends Controller
     {
         return view('productos.editProd', compact('producto'));
     }
-
-    // public function update(Request $request, Producto $producto)
-    // {
-    //     $request->validate([
-    //         'tipoproducto' => 'required|string',
-    //         'referencia' =>'required|string',
-    //         'descripcion' =>'required|string',
-    //         'alto' =>'required|string',
-    //         'ancho' =>'required|string',
-    //     ]);
-
-    //     $producto->update($request->all());
-
-    //     return redirect()->route('product.index')->with('success', 'Producto actualizado correctamente');
-    // }
-
 
     public function update(Request $request, Producto $producto)
     {
@@ -182,21 +148,39 @@ class ProductosController extends Controller
 
 
 
+    // public function destroy(Producto $producto)
+    // {
+    //     // Obtener la ruta del archivo PDF asociado al producto
+    //     $pdfPath = storage_path('app/public/' . $producto->plano);
+
+    //     // Eliminar el producto
+    //     $producto->delete();
+
+    //     // Si la ruta del archivo PDF existe, eliminarlo del sistema de archivos
+    //     if (file_exists($pdfPath)) {
+    //         unlink($pdfPath);
+    //     }
+
+    //     return response()->json(['success' => 'Producto eliminado satisfactoriamente']);
+    // }
     public function destroy(Producto $producto)
-    {
-        // Obtener la ruta del archivo PDF asociado al producto
+{
+    // Si el producto tiene un archivo PDF asociado, obtenemos la ruta del archivo y lo eliminamos si existe
+    if (!empty($producto->plano)) {
         $pdfPath = storage_path('app/public/' . $producto->plano);
-
-        // Eliminar el producto
-        $producto->delete();
-
-        // Si la ruta del archivo PDF existe, eliminarlo del sistema de archivos
         if (file_exists($pdfPath)) {
             unlink($pdfPath);
         }
-
-        return response()->json(['success' => 'Producto eliminado satisfactoriamente']);
     }
+
+    // Eliminar el producto de la base de datos
+    $producto->delete();
+
+    // Devolver una respuesta JSON indicando que el producto se eliminó correctamente
+    return response()->json(['success' => 'Producto eliminado satisfactoriamente']);
+}
+    
+
